refactor(protocol-designer): move deck setup folder to be in correct page
closes AUTH-674
