+++ protocol-designer/src/localization/en/protocol_overview.json
@@ -1,5 +1,3 @@
 {
-  "add_labware": "Add labware",
-  "edit": "Edit",
   "protocol_overview": "Protocol overview"
 }
