+++ protocol-designer/src/localization/en/starting_deck_state.json
@@ -1,3 +1,5 @@
 {
+  "add_labware": "Add labware",
+  "edit": "Edit",
   "starting_deck_state": "Starting deck state"
 }
